+++ pkgs/jnigen/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Breaking Change**([#1644](https://github.com/dart-lang/native/issues/1644)):
   Generate null-safe Dart bindings.
+- Simple AST for user Visitors
 
 ## 0.12.2